AP-3834: Do not persist the volume for bundled gems
This caused a problem running locally following
a `bundler` update. The cached version of the files
was not updated and resulted in the follwing error
during `docker compose up`

```shell
/usr/local/bundle/bin/bundle:25:in `load': cannot load such file -- /usr/local/bundle/gems/bundler-2.4.7/exe/bundle (LoadError)
```

This may also be true when any gem files or js pacakges
are updated so am removing the node_modules volume too.

It is possible to use `docker compose down` to allow them
to be updated by the next `build` or `up` but have chosen
to remove to avoid potentially confusing errors.