AP-3833: change name of secrets
We are intending to have multiple environments on cloud-platform

If we want to take advantage of the automatic rotation of k8s
cluster secrets we cannot use GHA environment secrets to differentiate
our k8s namespaces/environments. Instead we use the automatic naming
of secrets in combination with the ability provider by the serviceaccount
terraform module to specifying what names to use.