Fix docker compose run for assets
Without RAILS_SERVE_STATIC_FILES a production
like config on a local machine will not serve assets
from the /public folder. This is because
production instances expect to run on apache or nginx
(in our case) which will do this by default regardless.

We may not need to add this for an actual hosted environment
therefore, but if so, we can add it simply enough via deployment
pipeline, production config or the Dockerfile itself.