AP-3842: find users by subject id first
For first time logons using azure we want
to find authorised users by email and provider
only. This updates the subject id and name
based on the retrieved data from azure.

subsequent logins should use the subject id,
which is more secure as it cannot be (is more difficult to)
hijacked.