AP-3842: Add sign in flash message and ignore devise unused
Add a sign in success flash for consistency with sign out and
amend tests to exercise this.

Have also moved devise translations to locales structure we
are using for the app, normalized them and marked unused
translations as ignorable by the i18n-tasks gem spec purposes
(in case we want to amend them or devise module translations in the future).