AP-3963: Add non-JS file upload routes, controllers and forms
Upload and attaching of a file to a bulk submission object

Inline with progresive web app design the this should work
with no javascript.

Using form object to enable file validations/checking
BEFORE attaching. This is particualrly important for
eventual malware checking to occur before uploading
the actual file to s3.