AP-4043: change queue naming and configuration
To allow jobs to be enqueued on branch specific queues
for uat but non-branch specific queues for other environments.

UAT branches must only consume queues specific to the branch
as they share the same redis instance. If this is not done
then branch 1s jobs will be enqueued on a queue that branch
2's sidekiq processes can perform, introducing a race condition
that can result in confusing errors (for exampple, when the code
to perform the job does not existm, or database records on that
branch's database do not exist)