AP-4043: Move default queue naming enqueuing logic
Rather than using Worker.set(queue: queue_name).perform_async
to define which queue should process the job we can use
sidekiq_options in a more idiomatic way for clarity.

It may be possible to do the same for SubmissionQueueNameService
but this requires a use_case argument that is only available
in the perform action at present.