AP-4043: change interval algorithm to default
In testing a large bulk submission with details relating
to existing individual on the sandbox it was realised that
successful finds take quite a bit longer to return
a result - due to HMRC interface performing more requests
and HMRC API throttling those request.

A retry intervl of 5 resulted in no jobs being exhausted on
several occassions and so is something we could push to
production as a trial.