AP-4054: exhausted bulk_submission processing after 6 retries
We cannot keey checking for a bulk submission forever. This sets
the limit to 1 more than for the HMRC interface result worker
which means appoximately 20 minutes total wait time.

After that we mark the bulk submission as exhausted for debugging
and report to sentry.

A result_file will not be produced as a result - we could
write one anyway, and handle nil hmrc_interface_results when doing
so, if this becomes an issue.