AP-4158: Add step to install clamav client requirments
As well as having a "service" running th clamav image
used in production, we need to open the expected port
on that image so a client can communicate with it in
the test container/executor itself.

Note that not doing this can result in false positives
as the #virus_found? method returns true if there is
any error, not necessarily from virus existing. An additional
test will be added to ensure this does not occur.

The following error will return `#virus_found? true`
```
ERROR: Could not connect to clamd on localhost: Connection refused
```

local configuration to ensure the expected port, 3310, will be
added to the repos setup and readme