Remove use of LOCAL_CLAMAV
The bin/install_clamav_on_mac was altered to use TCP
meaning local clamav should work in the same way
as production clamav client - it will work using the
`clamd.container.conf` setting.

Therefore we do not need the LOCAL_CLAMAV command which uses
`--fdpass` (for improved speed?!).

```
--fdpass
Pass the file descriptor permissions to clamd. This is useful if clamd is
running as a different user as it is faster than streaming the file to clamd.
Only available if connected to clamd via local(unix) socket.
```