AP-6241: Use modsecurity for uat 503 and add add weekend test
Add day of week 503 error for Saturday (6) and Sunday (7).

Note that the documentation and searches state that a 0-6 day
week where saturday is 6 and sunday is 0. I have found that
this is not the case through trial and error, with 1-7 being the
weekday values with 1 being monday and sunday being 7.